+++ src/test/java/com/devcourse/ReviewRanger/user/service/UserServiceTest.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.devcourse.ReviewRanger.user.application.UserService;
 import com.devcourse.ReviewRanger.user.dto.JoinRequest;
 
 @Transactional
